+++ headers/helpers.h
@@ -13,7 +13,6 @@ void ResizeConsole(int width, int height);
 void colorTexto (int x);
 void endLines(int x);
 void iniciar_vector(int *vector, int filas);
-int sumar_dados(int vec[3], int CANT_DADOS);
 void obtener_maximo(int total, int &maximo, int i, int &jugadorMaximo);
 
 #endif // HELPERS_H
