+++ headers/juego.h
@@ -2,11 +2,11 @@
 #ifndef JUEGO_H
 #define JUEGO_H
 
-#include <Windows.h>
-#include <iostream>
-#include <string>
+// #include <Windows.h>
+// #include <iostream>
+// #include <string>
 #include "../juego.cpp"
-using namespace std;
+//using namespace std;
 
 void comenzar_juego(Jugadores *jugador);
