+++ headers/menu.h
@@ -2,18 +2,18 @@
 #ifndef MENU_H
 #define MENU_H
 
-#include <Windows.h>
-#include <iostream>
-#include<locale>
-#include <iomanip>
-#include <string>
-#include "rlutil.h"
-#include "jugadores.h"
-#include "dado.h"
-#include "helpers.h"
-#include "graficas.h"
+// #include <Windows.h>
+// #include <iostream>
+// #include<locale>
+// #include <iomanip>
+// #include <string>
+// #include "rlutil.h"
+// #include "jugadores.h"
+// #include "dado.h"
+// #include "helpers.h"
+// #include "graficas.h"
 #include "../menu.cpp"
-using namespace std;
+//using namespace std;
 
 void menuInicial();
