+++ headers/quien_empieza.h
@@ -2,7 +2,7 @@
 #ifndef QUIEN_EMPIEZA_H
 #define QUIEN_EMPIEZA_H
 
-#include "../quien_empieza.cpp"
+#include "../quien_empieza.cpp" 
 
 void quien_empieza(Jugadores *jugador, int &jugadorActual);
 void obtener_quien_empieza(Jugadores *jugador, int &jugadorActual, int jugadorDadoMaximo, int jugadorSumaMaxima);
