+++ main.cpp
@@ -16,5 +16,6 @@ int main (void){
   setlocale(LC_CTYPE, "Spanish"); // Para que se vea correctamente el texto en la consola
   ResizeConsole(850,550); // Cambiar0 tamano de consola
   menuInicial(); // Menu inicial
+
   return 0;
 }
